fix(contracts): fix cargo test failures across all contracts
- lazy_mint_erc721/test.rs: wrap two orphaned code blocks in proper
  test functions (allowlist_outsider_wrong_proof_returns_invalid_merkle_proof
  and allowlist_valid_proof_bad_sig_errors_not_allowlist); fix
  migrate_emits_migrated_event to use XDR-based events().iter() API
- lazy_mint_erc1155/test.rs: fix migrate_emits_migrated_event to use
  XDR ContractEventBody API instead of deprecated events.iter() tuple form
- collection_nft_erc721/test.rs: same migrate event fix; remove usage of
  Symbol::try_from_val which is not available in soroban-sdk v25
- collection_nft_erc1155/test.rs: same migrate event fix
- launchpad/types.rs: change PreflightResult.errors from Vec<Error> to
  Vec<u32> to avoid SorobanArbitrary bound that contracterror enums cannot
  satisfy in test builds
- launchpad/contract.rs: add errors_to_u32() helper; update all four
  preflight_deploy_* functions to convert Vec<Error> -> Vec<u32>
- launchpad/test.rs: update preflight error assertions from
  contains(&Error::X) to contains(&(Error::X as u32))

+++ contracts/launchpad/src/test.rs
@@ -2344,7 +2344,7 @@ fn preflight_normal_721_predicts_the_deployed_address() {
         &salt,
     );
     assert_eq!(preflight_after.predicted_address, deployed);
-    assert!(preflight_after.errors.contains(&Error::DuplicateSalt));
+    assert!(preflight_after.errors.contains(&(Error::DuplicateSalt as u32)));
 
     assert_eq!(
         client.try_deploy_normal_721(
@@ -2386,12 +2386,12 @@ fn preflight_normal_721_reports_every_error_deploy_would_raise() {
         &salt,
     );
 
-    assert!(preflight.errors.contains(&Error::ContractPaused));
-    assert!(preflight.errors.contains(&Error::EmptyName));
-    assert!(preflight.errors.contains(&Error::EmptySymbol));
-    assert!(preflight.errors.contains(&Error::InvalidMaxSupply));
-    assert!(preflight.errors.contains(&Error::InvalidRoyaltyBps));
-    assert!(preflight.errors.contains(&Error::InvalidFeeBps));
+    assert!(preflight.errors.contains(&(Error::ContractPaused as u32)));
+    assert!(preflight.errors.contains(&(Error::EmptyName as u32)));
+    assert!(preflight.errors.contains(&(Error::EmptySymbol as u32)));
+    assert!(preflight.errors.contains(&(Error::InvalidMaxSupply as u32)));
+    assert!(preflight.errors.contains(&(Error::InvalidRoyaltyBps as u32)));
+    assert!(preflight.errors.contains(&(Error::InvalidFeeBps as u32)));
 
     // The real deploy call rejects on the first violation it checks
     // (contract-paused) — proving preflight is a superset, never blind to a
@@ -2434,7 +2434,7 @@ fn preflight_flags_insufficient_balance_for_the_flat_fee() {
         &salt,
     );
 
-    assert!(preflight.errors.contains(&Error::InsufficientFee));
+    assert!(preflight.errors.contains(&(Error::InsufficientFee as u32)));
     assert_eq!(preflight.required_fee, 1_000i128);
 
     // The real deploy call fails too — it attempts the token transfer, which
